login module and refacto
